[修改] 大框架修改

+++ project/setup.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# -----------------------------------------------
-# Copyright 2023 for Fosun. All Rights Reserved.
-# -----------------------------------------------
+# ---------------------------------------------
+# Copyright 2015 for Zen. All Rights Reserved.
+# ---------------------------------------------
 
 import os
 import sys
@@ -10,7 +10,6 @@ import json
 from pathlib import Path
 from pathspec import PathSpec
 from setuptools import setup
-from termcolor import cprint
 from Cython.Build import cythonize, build_ext
 
 
@@ -32,21 +31,19 @@ class Compile(object):
     operation = sys.argv[1] if len(sys.argv) > 1 else 'check'
     os.system(f"find {ProjectPath} -name '__pycache__' -o -name '.DS_Store' | xargs rm -rfv")
 
-    def __init__(self):
-        if self.operation == 'check':
-            self.check()
-        elif self.operation == 'build':
-            self.build()
-        elif self.operation == 'clean':
-            self.clean()
+    def __init__(self, debug=False):
+        if self.operation == 'build':
+            self.build(debug)
         elif self.operation == 'rollback':
             self.rollback()
+        elif self.operation == 'check':
+            self.check()
         else:
-            raise IOError("使用方式：python setup.py [选择: check | build | clean | rollback ]")
+            raise IOError("使用方式：python setup.py [选择: build | rollback | check ]")
 
     @staticmethod
-    def build():
-        os.system(f"rm -rfv {BackupPath}/* && mkdir -p {BackupPath} && cp -rfv {AlgorithmsPath} {BackupPath}")
+    def build(debug):
+        os.system(f"mkdir -p {BackupPath} && cp -rfv {AlgorithmsPath} {BackupPath}")
 
         extensions = []
         for rootpath, dirnames, filenames in os.walk(AlgorithmsPath):
@@ -65,22 +62,18 @@ class Compile(object):
             ext_modules=modules,
             cmdclass={'build_ext': build_ext},
         )
-        cprint(f"编译已完成。", color='red')
+        print(f"文件编译已完成。")
 
         os.system(f"cp -rfv {BuildPath}/lib*/algorithms {ProjectPath}")
         os.system(f"find {AlgorithmsPath}/* -name '*.c' -o -name '*.py' | egrep -v '__init__.py' | xargs rm -rfv")
-        cprint(f"安装已完成。", color='red')
+        os.system(f"rm -rfv {BackupPath} {BuildPath}") if not debug else None
+        print(f"安装清理已完成。")
 
     @staticmethod
     def rollback():
         os.system(f"rm -rfv {AlgorithmsPath}/*")
         os.system(f"cp -rfv {BackupPath}/algorithms {ProjectPath}")
-        cprint(f"回滚已完成。", color='red')
-
-    @staticmethod
-    def clean():
-        os.system(f"rm -rfv {BackupPath} {BuildPath}")
-        cprint(f"已删除所有源码！", color='red')
+        print(f"回滚已完成。")
 
     @staticmethod
     def check():
@@ -89,7 +82,6 @@ class Compile(object):
             for filename in filenames:
                 if not Exclude(f"{rootpath}/{filename}"):
                     extensions.append(f"{rootpath}/{filename}")
-
         print(json.dumps(extensions, indent=4, ensure_ascii=False))
